Feature/plot orphaning

+++ src/main/java/com/atherys/towns/facade/TownFacade.java
@@ -217,6 +217,10 @@ public class TownFacade implements EconomyFacade {
             throw new TownsCommandException("You cannot unclaim your last remaining plot.");
         }
 
+        if (townService.checkPlotRemovalCreatesOrphans(town, plot)) {
+            throw new TownsCommandException("You cannot unclaim a plot that would result in orphaned plots.");
+        }
+
         townService.removePlotFromTown(town, plot);
         townsMsg.info(source, "Plot abandoned.");
     }
@@ -225,7 +229,6 @@ public class TownFacade implements EconomyFacade {
         PlotSelection selection = plotSelectionFacade.getValidPlayerPlotSelection(source);
 
         Town town = getPlayerTown(source);
-
         Plot plot = plotService.createPlotFromSelection(selection);
 
         if (townService.getTownSize(town) + plotService.getPlotArea(plot) > town.getMaxSize()) {
